CLI-837 Move git related code to /core/host/git
