+++ src/commands/integrate/_common/preflight-summary.ts
@@ -21,12 +21,12 @@
 // Preflight summaries shown at the start of integrate commands, before install.
 
 import { CommandFailedError } from '@/core/command-error.ts';
+import { GitRepo } from '@/core/host/git/hooks.ts';
 import type { DiscoveredProject } from '@/core/project-info.ts';
 import { SonarQubeClient } from '@/core/server/client.ts';
 import type { PhaseItem, StepStatus } from '@/core/ui';
 import { info, outro, phase, phaseItem, text } from '@/core/ui';
 
-import { GitRepo } from '../../_common/git-repo.ts';
 import { checkTokenStatus, type TokenCheckResult, type TokenStatus } from '../../_common/token.ts';
 
 export interface AgentPreflightSummaryOptions {
