+++ src/commands/integrate/_common/registry/install-integration.ts
@@ -20,7 +20,7 @@
 
 import { CommandFailedError } from '@/core/command-error.ts';
 import type { ResolvedAuth } from '@/core/host/auth-resolver.ts';
-import { findGitRoot } from '@/core/host/discover-git-repo.ts';
+import { findGitRoot } from '@/core/host/git/discover.ts';
 import logger from '@/core/observability/logger.ts';
 import type {
   CliState,
