+++ src/core/project-info.ts
@@ -25,7 +25,7 @@ import { basename, join } from 'node:path';
 
 import type { ResolvedAuth } from '@/core/host/auth-resolver.ts';
 import { resolveAuth } from '@/core/host/auth-resolver.ts';
-import { findGitRoot, getGitRemote } from '@/core/host/discover-git-repo.ts';
+import { findGitRoot, getGitRemote } from '@/core/host/git/discover.ts';
 import {
   discoverProjectKeyByGitRemote,
   GIT_REMOTE_BINDING_SOURCE,
